+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 12.41.0
+
+* `[usage]` Added helpers for generating a string representation of `Command` and `Option`
+* `[fsutil]` Improved errors description for `ValidatePerms`
+
 ### 12.40.0
 
 * `[ansi]` Added methods for working with byte slices